+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `fetch_all_key_vaults` fn
 - Add `KeyVault` and `KeyVaultId` and `KeyVaultProperties` and `KeyVaultName` types
 - Remove storage account duplicate properties already exposed by the id
+- Rename PIM role assignment stuff to `GovernanceRoleAssignment`
 
 # v0.25.0